added: idxtostr, strtoidx; new target: IDX_UKWNC99.
I felt that `strtoidx` was necessary, since strtoull returns an
`unsigned long long` and an `idx_t` isn't guaranteed to be an `ull`.
On top of that I'm not 100% Microsoft support `strtoull` on all their
systems.

`idxtostr` is probably a little superfluous but I added it for
completion-sake.

The IDX_UKWNC99 target now became necessary since `strtoull` is only
available on systems with C99+.